+++ src/libraries/Common/src/System/Buffers/Text/Base64Decoder.cs
@@ -477,22 +477,6 @@ namespace System.Buffers.Text
             destBytes = dest;
         }
 
-        // This can be replaced once https://github.com/dotnet/runtime/issues/63331 is implemented.
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static Vector128<byte> SimdShuffle(Vector128<byte> left, Vector128<byte> right, Vector128<byte> mask8F)
-        {
-            Debug.Assert((Ssse3.IsSupported || AdvSimd.Arm64.IsSupported) && BitConverter.IsLittleEndian);
-
-            if (Ssse3.IsSupported)
-            {
-                return Ssse3.Shuffle(left, right);
-            }
-            else
-            {
-                return AdvSimd.Arm64.VectorTableLookup(left, Vector128.BitwiseAnd(right, mask8F));
-            }
-        }
-
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static unsafe void Vector128Decode(ref byte* srcBytes, ref byte* destBytes, byte* srcEnd, int sourceLength, int destLength, byte* srcStart, byte* destStart)
         {
